Extracted TableKeyBinder and register IdentifierBinding with BindingState for later resolution
